Fi 2718 _include search parameter fix. (#12)
# Patch Implementation Guide

Patch capability statement to have correct "_include" syntax. Was "searchParam". Should have been "sourceType:searchParam".
The Capability Statement required updates to include the correct _include syntax.

I wanted to ensure it was clear that we were doing a patch to the implementation guide and why we were doing the patch, so I did not change the Implementation Guide directly. Instead, I put the updated Capability Statement in a folder called "package_supersede". Any files in this folder will supersede files in the Implementation Guide (package.tgz).

There was a similar existing patch for r4_search-parameters. I moved this file into "package_supplement", because we do not want this file to supersede existing files, we just want it to add in missing information.

The load order determines what files "supersede" and what files "supplement".

IGResources accesses resources by grabbing the first matching resource it finds for a request. Files that are loaded first take precedence.

So the load order is:

    package_supersede
    package
    package_supplement